Implement lsub for mbox.
The IMAP lsub command has this weird requirement associated with the %
wildcard as a trailing character.  Basically, you can list a mail folder
that isn't in the subscribed list if it holds items that are subscribed.  I
implement this by using a regex mechanism similar to the one I use for list,
but I modify the regex slightly for lsub (I don't require that it match the
whole string if the last character in the specification is a %) and I check
the string that was matched against the line and mark it as specified in
RFC 3501 if they don't match.

darcs-hash:20070512143344-1210e-26631312d9730dcf59d2ad52c942d32f0635daff