Fix a buffer overflow in SubscribeHandlerExecute
It turns out that the responseText buffer was being overflowed because
strncat didn't work the way that I expected and because I neglected to leave
room for the terminator in the buffers.  I also made the buffer bigger
because I want to return a completely descriptive error message to the
client.

darcs-hash:20070512161254-1210e-29a5f1c6545ab2082804a706ac36490fca1f310c