Make GetFQDN not static
This is important for two reasons.  First, the calculation of the FQDN may
be expensive so I want to be able to cache the results in the server class.
Second, I might have several instances of the server running (say to have
a set of virtual servers running on different addresses) and need to pull
the data from the server associated with the given session rather than one
that just happens to be for one of the servers in question.

darcs-hash:20070518173805-1210e-04709b6790db91ce8ad9d9b5d3e0f617c791926d