Rework the DateTime class
There are three major things I've done.  First, I've added a "convert to
string" method so I can use the date and time.  Second, I've redefined the
exception such that it just means "invalid" so it makes sense to throw it
from the print function.  Third, if the available data is too short on the
input value, I throw the exception, which was a case I simply forgot about
when I was creating the class last night.

darcs-hash:20070527040134-1210e-dbc44705d3b587a4801450e31fc08575fd8acaf5