Add a "done appending" method to the mail store.
Right now, I'm thinking that there are going to be two major approaches to
appending messages.  The first is to create an object in and of itself that
can be modified where you you don't know (nor care) that all the updates are
finished, and the second is to modify a part of another object and to stop
modifying it when you're done.  The first way is how I did it with the mail
store I invented for Simdesk, the second way is more appropriate for mbox.

darcs-hash:20070527040508-1210e-344914cc0b166d4d80a0d6c4f8e25e16219fa52d