Storing whether or not a mailbox is open doesn't help much.
Instead, I need to store which mailbox is open because not only do I need to
worry about whether or not the mailbox I'm appending to is the currently
open mailbox, but eventually I'm going to issue commands that are expected
to update the currently open mail file so I have have some idea what that
is.  If I use a pointer to a string to hold the mailbox name, I can use
whether or not that pointer is NULL to indicate whether any mailbox is open
and the value pointed to can hold the name of the mailbox.

Note that I use the IMAP name as the mailbox, not the path to the file.  I'm
not sure why I'm doing it this way, it just seems right.

darcs-hash:20070531011136-1210e-2bc40a88a5a33bb09af7ac6d83cb8712949d616a