Fix the defect where copying the torture-test message corrupts the torture-test file.
There are two fixes.  First, I need to not mark the original file is dirty
if I'm copying to a different file.  I do that by checking the name of the
open file against the name of the file to append to.  Second, I need to not
just sort of stop in the middle of processing the file when rewriting long
files.

I hadn't seen the second issue because I've been working with small files.

darcs-hash:20070704032338-1210e-13800f38d9f8821ccc81b06596f0e95d299f2bd3