wrap the append in a lock/unlock pair
Ignore-this: 15d4bf2f6032d6ad87bc28a9f92bd1f5

Any multistep process, and an append process where it creates the message, then
writes bits of the message to the message, and then closes the message, should
be done in a single lock.  So I wrap the calls in a lock/unlock pair.  The main
problem with this approach is that I didn't have any mechanism for setting a lock
on a mailbox that wasn't currently selected by that session.  However, append
happens to an unselected mailbox.  So, I had to add a mechanism for setting a
lock on a mailbox that isn't currently selected.

darcs-hash:20110613152151-1210e-ced87cf95dec1f09c462f520b57e9455aa408c34