Major refactoring.
Ignore-this: 1f27de7f7410f7ec8ebdbd2d2b4e7cae

The goal of this was to put the bulk of the complexity into individual handler
classes and remove it (largely) from the session class.  This should make the
logic clearer and make it a lot easier to write unit tests for the commands.

I also redid a lot about the parsing, so I bet I added a number of bugs to the
code.  It'll be a bear to test all of this, making me wish I'd already written
those aforementioned unit tests.

darcs-hash:20131129234730-1210e-83713f6e43a79148ad03d65b2b58bceaeeaaf523