Make it easier to unit test the handlers and to mock the calls used by the unimplemented command handler.
The unit testing of the handlers is made easier by checking for NULL pointers
before calling the driver to emit a message.  That allows me to create the
handler by constructing a session with NULL pointers for all of the constructor
arguments.  That means I don't have to figure out how to build a server object
or a driver object or whatever just to test handlers.

The mocking for the unimplemented handler is made possible by declaring
the responseText method virtual.